Move IStackInventory tests to StackInventory
