--- tests/Containers/Interfaces/IStackInventory/IStackInventoryTests.AddItem.cs
@@ -1,16 +1,0 @@
-﻿namespace TheChest.Inventories.Tests.Containers.Interfaces
-{
-    public partial class IStackInventoryTests<T>
-    {
-        [Test]
-        public void AddItem_EmptyInventory_ReturnsTrue()
-        {
-            var item = this.itemFactory.CreateDefault();
-            var inventory = this.inventoryFactory.EmptyContainer();
-
-            var result = inventory.Add(item);
-
-            Assert.That(result, Is.True);
-        }
-    }
-}
