--- tests/Containers/Interfaces/IStackInventory/IStackInventoryTests.CanReplace.cs
@@ -1,9 +1,0 @@
-using TheChest.Tests.Common.Attributes;
-
-﻿namespace TheChest.Inventories.Tests.Containers.Interfaces
-{
-    public partial class IStackInventoryTests<T>
-    {
-
-    }
-}
