--- tests/Containers/Interfaces/IStackInventory/IStackInventoryTests.Move.cs
@@ -1,6 +1,0 @@
-﻿namespace TheChest.Inventories.Tests.Containers.Interfaces
-{
-    public partial class IStackInventoryTests<T>
-    {
-    }
-}
