+++ tests/Containers/StackInventory/StackInventoryTests.AddItem.cs
@@ -206,5 +206,17 @@ namespace TheChest.Inventories.Tests.Containers.StackInventory
             
             Assert.That(raised, Is.True);
         }
+
+
+        [Test]
+        public void AddItem_EmptyInventory_ReturnsTrue()
+        {
+            var item = this.itemFactory.CreateDefault();
+            var inventory = this.inventoryFactory.EmptyContainer();
+
+            var result = inventory.Add(item);
+
+            Assert.That(result, Is.True);
+        }
     }
 }
