+++ tests/Containers/StackInventory/StackInventoryTests.AddItems.cs
@@ -268,5 +268,18 @@ namespace TheChest.Inventories.Tests.Containers.StackInventory
                 Assert.That(inventory.GetItems(1), Is.EqualTo(items.Skip(stackSize)));
             });
         }
+
+
+        [Test]
+        public void AddItems_EmptyInventory_ReturnsEmptyArray()
+        {
+            var size = this.random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
+            var items = this.itemFactory.CreateMany(size);
+            var inventory = this.inventoryFactory.EmptyContainer();
+
+            var result = inventory.Add(items);
+
+            Assert.That(result, Is.Empty);
+        }
     }
 }
