+++ tests/Containers/StackInventory/StackInventoryTests.GetItem.cs
@@ -1,4 +1,5 @@
-﻿using TheChest.Tests.Common.Extensions.Containers;
+﻿using TheChest.Tests.Common.Attributes;
+using TheChest.Tests.Common.Extensions.Containers;
 
 namespace TheChest.Inventories.Tests.Containers.StackInventory
 {
@@ -77,6 +78,74 @@ namespace TheChest.Inventories.Tests.Containers.StackInventory
             
             var item = this.itemFactory.CreateDefault();
             inventory.Get(item);
+        }
+
+
+        [Test]
+        [IgnoreIfValueType]
+        public void GetItem_EmptyInventory_ReturnsNull()
+        {
+            var item = this.itemFactory.CreateDefault();
+            var inventory = this.inventoryFactory.EmptyContainer();
+
+            var foundItem = inventory.Get(item);
+            
+            Assert.That(foundItem, Is.Null);
+        }
+
+        [Test]
+        [IgnoreIfReferenceType]
+        public void GetItem_EmptyInventoryValueType_ReturnsDefault()
+        {
+            var item = this.itemFactory.CreateDefault();
+            var inventory = this.inventoryFactory.EmptyContainer();
+
+            var foundItem = inventory.Get(item);
+
+            Assert.That(foundItem, Is.EqualTo(default(T)));
+        }
+
+        [Test]
+        public void GetItem_InventoryWithItems_ReturnsItem()
+        {
+            var size = this.random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
+            var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
+            var slotItem = this.itemFactory.CreateRandom();
+            var inventory = this.inventoryFactory.FullContainer(size, stackSize, slotItem);
+
+            var item = inventory.Get(slotItem);
+
+            Assert.That(item, Is.EqualTo(slotItem));
+        }
+
+        [Test]
+        [IgnoreIfValueType]
+        public void GetItem_InventoryWithDifferentItems_ReturnsNull()
+        {
+            var size = this.random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
+            var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
+            var slotItem = this.itemFactory.CreateRandom();
+            var inventory = this.inventoryFactory.FullContainer(size, stackSize, slotItem);
+
+            var item = this.itemFactory.CreateDefault();
+            var result = inventory.Get(item);
+
+            Assert.That(result, Is.Null);
+        }
+
+        [Test]
+        [IgnoreIfReferenceType]
+        public void GetItem_InventoryWithDifferentItemsValueType_ReturnsDefault()
+        {
+            var size = this.random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
+            var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
+            var slotItem = this.itemFactory.CreateRandom();
+            var inventory = this.inventoryFactory.FullContainer(size, stackSize, slotItem);
+
+            var item = this.itemFactory.CreateDefault();
+            var result = inventory.Get(item);
+
+            Assert.That(result, Is.EqualTo(default(T)));
         }
     }
 }
