+++ tests/Containers/StackInventory/StackInventoryTests.GetItems.cs
@@ -104,5 +104,55 @@ namespace TheChest.Inventories.Tests.Containers.StackInventory
             };
             inventory.Get(item, 100);
         }
+
+
+        [Test]
+        public void GetItems_ItemNotFound_ReturnsEmptyArray()
+        {
+            var item = this.itemFactory.CreateDefault();
+            var inventory = this.inventoryFactory.EmptyContainer();
+
+            var items = inventory.Get(item, 1);
+
+            Assert.That(items, Is.Empty);
+        }
+
+        [Test]
+        public void GetItems_ItemsFound_ReturnsItems()
+        {
+            var size = this.random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
+            var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
+            var items = this.itemFactory.CreateManyRandom(size / 2);
+            var inventoryItems = this.itemFactory
+                .CreateManyRandom(size / 2)
+                .ToList();
+            inventoryItems.AddRange(items);
+
+            var inventory = this.inventoryFactory.ShuffledItemsContainer(size, stackSize, inventoryItems.ToArray());
+
+            var expectedItem = items[0];
+            var amount = this.random.Next(1, stackSize);
+            var result = inventory.Get(expectedItem, amount);
+
+            Assert.That(result, Has.Length.EqualTo(amount));
+            Assert.That(result, Has.All.EqualTo(expectedItem));
+        }
+
+        [Test]
+        public void GetItems_AmoutBiggerThanItemsInInventory_ReturnsAllItemsFound()
+        {
+            var size = this.random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
+            var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
+            var item = this.itemFactory.CreateDefault();
+            var inventoryItems = this.itemFactory.CreateManyRandom(size - 1)
+                .Append(item)
+                .ToList();
+
+            var inventory = this.inventoryFactory.ShuffledItemsContainer(size, stackSize, inventoryItems.ToArray());
+            var result = inventory.Get(item, 100);
+
+            Assert.That(result, Has.Length.EqualTo(stackSize));
+            Assert.That(result, Has.All.EqualTo(item));
+        }
     }
 }
